Rename the Lumen GraphQL controller
